backup/requirements.txt
python-dotenv
azure-cosmos
